More review comments: valid_origin and ListingType::from
* Added a From trait for ListingType to convert from ArgMatches.
* Caught a missing error case where I wasn't handling bad origin names
