+++ components/hab/src/main.rs
@@ -59,7 +59,6 @@ use hcore::crypto::keys::PairType;
 use hcore::crypto::{default_cache_key_path, init, BoxKeyPair, SigKeyPair};
 use hcore::env as henv;
 use hcore::fs::{cache_analytics_path, cache_artifact_path, cache_key_path};
-use hcore::package::ident;
 use hcore::package::PackageIdent;
 
 use hcore::service::ServiceGroup;
@@ -688,7 +687,7 @@ fn sub_pkg_path(m: &ArgMatches) -> Result<()> {
 }
 
 fn sub_pkg_list(m: &ArgMatches) -> Result<()> {
-    let listing_type = ident_or_origin(&m);
+    let listing_type = ListingType::from(m);
 
     command::pkg::list::start(&listing_type, &*FS_ROOT)
 }
@@ -1319,26 +1318,6 @@ fn org_param_or_env(m: &ArgMatches) -> Result<String> {
             Ok(v) => Ok(v),
             Err(_) => return Err(Error::CryptoCLI("No organization specified".to_string())),
         },
-    }
-}
-
-/// Check to see if a package idenifier or
-///
-fn ident_or_origin(m: &ArgMatches) -> ListingType {
-    if m.is_present("ALL") {
-        return ListingType::AllPackages;
-    }
-    if m.is_present("ORIGIN") {
-        let origin = m.value_of("ORIGIN").unwrap(); // Required by clap
-        if ident::is_valid_origin_name(&origin) {
-            return ListingType::Origin(origin.to_string());
-        }
-    }
-
-    let p = m.value_of("PKG_IDENT").unwrap(); // Required by clap
-    match PackageIdent::from_str(&p) {
-        Ok(ident) => ListingType::Ident(ident),
-        Err(_) => unreachable!("We've already validated PackageIdent {}", &p),
     }
 }
